refactor(webapp): extract shared meeting modules from modal features
Move defaults, styles, participants, commands, service logic, and submit
helpers into Components/Meeting/shared so Meet Now and Schedule Meeting
no longer depend on each other's internal modules.
