+++ apps/webapp/src/script/components/Meeting/ScheduleMeetingModal/scheduleMeetingTypes.ts
@@ -34,14 +34,6 @@ export interface ScheduleMeetingFormState {
   participantsFilter: string;
 }
 
-export type ScheduleMeetingCommand = {
-  title: string;
-  start: Date;
-  end: Date;
-  recurrence: ScheduleMeetingRecurrenceOption;
-  selectedUsers: User[];
-};
-
 export type ScheduleMeetingFormErrorKey =
   | 'meetings.scheduleModal.error.titleRequired'
   | 'meetings.scheduleModal.error.endBeforeStart'
